refactor : 로또 티켓을 ball 패키지로 이동
