+++ src/main/java/lotto/console/domain/user/User.java
@@ -6,7 +6,7 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 import lotto.console.domain.ball.WinningBalls;
-import lotto.console.domain.ticket.LottoTicket;
+import lotto.console.domain.ball.LottoTicket;
 
 public class User {
