+++ src/test/java/lotto/console/domain/user/RankTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import lotto.console.converter.Converter;
 import lotto.console.domain.ball.BallFactory;
 import lotto.console.domain.ball.WinningBalls;
-import lotto.console.domain.ticket.LottoTicket;
+import lotto.console.domain.ball.LottoTicket;
 
 public class RankTest {
